+++ src/app/app.module.ts
@@ -10,17 +10,20 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCardModule } from '@angular/material/card';
+import { MatTabsModule } from '@angular/material/tabs';
 
 import { TimerPageComponent } from './timer-page/timer-page.component';
 import { MainPageComponent } from './main-page/main-page.component';
 import { TimerComponent } from './timer/timer.component';
+import { TimerSpinnerComponent } from './timer-spinner/timer-spinner.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     TimerPageComponent,
     MainPageComponent,
-    TimerComponent
+    TimerComponent,
+    TimerSpinnerComponent
   ],
   imports: [
     FormsModule,
@@ -31,6 +34,7 @@ import { TimerComponent } from './timer/timer.component';
     MatToolbarModule,
     MatIconModule,
     MatCardModule,
+    MatTabsModule,
   ],
   providers: [],
   bootstrap: [AppComponent]