+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - improve style
   - change sound
   - sound last until click stop button or other mechanism
+  - improve spinner
 
 - add instructions
 - timer series
@@ -27,6 +28,7 @@
   2. change project name
   3. modify style
   4. change Copyright
+  5. add timer-spinner
 - 20210323 adjust timer and add time's up sound
   1. adjust timer
   2. add time's up sound